// src/types/youtube.d.ts
interface LiveStreamabilityRenderer {
    offlineSlate: {
        liveStreamOfflineSlateRenderer: {
            scheduledStartTime: string;
        }
    }
}

interface SubMenuItem {
    continuation: {
        reloadContinuationData: {
            continuation: string;
        }
    }
}

interface ViewSelector {
    sortFilterSubMenuRenderer: {
        subMenuItems: SubMenuItem[];
    }
}

export type VideoDetails = z.infer<typeof VideoDetailsSchema>;

export type ChannelMetadata = z.infer<typeof ChannelMetadataSchema>;